Pass Stage 5
Implement timing attack for password hacking

In order to improve the effectiveness of the password hacking method, a timing attack has been implemented. Now the 'Response' class includes an 'elapsedTime' property to measure the time it takes for a response to be returned. Besides, a new static method 'getByResponse' was added to 'ResponseType' class to utilize this new property. Adjustments were made on associated classes accordingly.